Move objectinstance and Object base class into instance.c/.h
