+++ src/builtin/builtin.c
@@ -294,6 +294,7 @@ void builtin_initialize(void) {
     /* Initialize builtin classes and functions */
     string_initialize(); // Must be initialized early as much depends on this
     function_initialize();
+    instance_initialize();
     
     object_initialize(); // This has been carefully placed to ensure
     
@@ -340,5 +341,8 @@ void builtin_finalize(void) {
     closure_finalize();
     list_finalize();
     dict_finalize();
+    invocation_finalize();
+    instance_finalize();
+    function_finalize();
     string_finalize();
 }
